Remvoving const module
