--- src/constants.js
@@ -1,39 +1,0 @@
-const KeyCodes = {
-  ESC: 27
-};
-
-const OffersSettings = {
-  MAX: 3
-};
-
-const SuccessStatusCodes = {
-  MIN: 200,
-  MAX: 300
-};
-
-const ANIMATION = {
-  timeout: 600,
-  duration: 0.6
-};
-
-const Method = {
-  GET: `GET`,
-  POST: `POST`,
-  PUT: `PUT`,
-  DELETE: `DELETE`
-};
-
-const AUTHORIZATION = `Basic eo0w590ik299a=asd`;
-
-const END_POINT = `https://es8-demo-srv.appspot.com/big-trip`;
-
-const BAR_HEIGHT = 55;
-
-export {KeyCodes};
-export {OffersSettings};
-export {SuccessStatusCodes};
-export {ANIMATION};
-export {AUTHORIZATION};
-export {END_POINT};
-export {BAR_HEIGHT};
-export {Method};
